+++ src/main/java/org/jetbrains/bsp/bazel/server/bep/utils/BUILD
@@ -1,0 +1,15 @@
+load("@rules_java//java:defs.bzl", "java_library")
+
+java_library(
+    name = "utils",
+    srcs = glob(["*.java"]),
+    visibility = ["//visibility:public"],
+    deps = [
+        "//src/main/java/org/jetbrains/bsp/bazel/commons",
+        "@io_bazel//src/main/protobuf:analysis_java_proto",
+        "@maven//:ch_epfl_scala_bsp4j",
+        "@maven//:com_google_guava_guava",
+        "@maven//:org_apache_logging_log4j_log4j_api",
+        "@maven//:org_apache_logging_log4j_log4j_core",
+    ],
+)
